feat(rpc): disallow deps into rpc client from outside
For this, common converters are extracted from rpc serializers.
